Fix authorization in proxy when password enabled
The client-side fetch API, by default, appends the correct authorization (which includes the username and password) to any requests. However, the code was overwriting it with API keys.

To fix this, I added a new header for the proxy called `X-REAL-AUTHORIZATION` which will proxy the API key through, while not overwriting the user/password authorization needed for the server to accept the proxy request.